+++ NEWS.md
@@ -1,4 +1,6 @@
 # pathfindR (development version)
+## Minor Changes
+- added the `disable_parallel` argument in `active_snw_enrichment_wrapper()` to be able to disable parallel runs via `foreach` 
 
 # pathfindR 2.1.0
 ## Minor Changes